requirements.txt
pytorch==2.3.1
torchvision==0.18.1
numpy==1.26.4
matplotlib==3.10.1
scikit-learn==1.6.1
